+++ crates/backend/src/http/auth/mod.rs
@@ -9,7 +9,7 @@ use actix_identity::Identity;
 use crate::database::{table, Database};
 
 
-pub fn get_auth_value(identity: &Identity) -> Option<i64> {
+pub fn get_auth_value(identity: &Identity) -> Option<usize> {
 	let ident = identity.identity()?;
 	serde_json::from_str(&ident).ok()
 }
@@ -19,6 +19,6 @@ pub fn get_auth_member(identity: &Identity, db: &Database) -> Option<table::Memb
 	db.get_member_by_id(id).ok().flatten()
 }
 
-pub fn remember_member_auth(member_id: i64, identity: &Identity) {
+pub fn remember_member_auth(member_id: usize, identity: &Identity) {
 	identity.remember(member_id.to_string());
 }
