+++ crates/frontend/src/components/reader/mod.rs
@@ -571,7 +571,7 @@ fn create_iframe() -> HtmlIFrameElement {
 		.unwrap()
 }
 
-fn generate_pages(book_dimensions: Option<(i32, i32)>, book_id: i64, chapter: Chapter, scope: Scope<Reader>) -> ChapterContents {
+fn generate_pages(book_dimensions: Option<(i32, i32)>, book_id: usize, chapter: Chapter, scope: Scope<Reader>) -> ChapterContents {
 	let iframe = create_iframe();
 
 	iframe.set_attribute(
